siapen-backend-development/pyproject.toml
[tool.poetry]
name = "p"
version = "0.1.0"
description = ""
authors = ["Fedorov <atiliogoulart@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
drf-yasg = "^1.20.0"
django-cors-headers = "3.6.0"
django-filter = "2.4.0"
djangorestframework = "3.12.2"
djangorestframework-simplejwt = "4.6.0"
Pillow = "^8.3.1"
python-decouple = "3.3"
psycopg2-binary = "2.8.6"
Unidecode = "1.1.2"
coreapi = "2.3.3"
PyYAML = "5.3.1"
requests = "^2.26.0"
coverage = "5.3.1"
image-helpers = "0.1.4"
validate-docbr = "1.8.2"
drf-flex-fields = "0.8.9"
django-ordered-model = "3.4.3"
python-dateutil = "2.8.1"
workalendar = "15.4.0"
drf-api-tracking = "1.8.0"
pre-commit = "^2.14.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
